benches: Add `r1cs` and `inner_product_proof` benchmarks
